refactor: convert controlUtils to TypeScript (2 of 2) (#13520)
