+++ superset-frontend/package.json
@@ -65,34 +65,34 @@
     "@babel/runtime-corejs3": "^7.12.5",
     "@data-ui/sparkline": "^0.0.84",
     "@emotion/core": "^10.0.35",
-    "@superset-ui/chart-controls": "^0.17.15",
+    "@superset-ui/chart-controls": "^0.17.17",
     "@superset-ui/core": "^0.17.15",
-    "@superset-ui/legacy-plugin-chart-calendar": "^0.17.15",
-    "@superset-ui/legacy-plugin-chart-chord": "^0.17.15",
-    "@superset-ui/legacy-plugin-chart-country-map": "^0.17.15",
-    "@superset-ui/legacy-plugin-chart-event-flow": "^0.17.15",
-    "@superset-ui/legacy-plugin-chart-force-directed": "^0.17.15",
-    "@superset-ui/legacy-plugin-chart-heatmap": "^0.17.15",
-    "@superset-ui/legacy-plugin-chart-histogram": "^0.17.15",
-    "@superset-ui/legacy-plugin-chart-horizon": "^0.17.15",
-    "@superset-ui/legacy-plugin-chart-map-box": "^0.17.15",
-    "@superset-ui/legacy-plugin-chart-paired-t-test": "^0.17.15",
-    "@superset-ui/legacy-plugin-chart-parallel-coordinates": "^0.17.15",
-    "@superset-ui/legacy-plugin-chart-partition": "^0.17.15",
-    "@superset-ui/legacy-plugin-chart-pivot-table": "^0.17.15",
-    "@superset-ui/legacy-plugin-chart-rose": "^0.17.15",
-    "@superset-ui/legacy-plugin-chart-sankey": "^0.17.15",
-    "@superset-ui/legacy-plugin-chart-sankey-loop": "^0.17.15",
-    "@superset-ui/legacy-plugin-chart-sunburst": "^0.17.15",
-    "@superset-ui/legacy-plugin-chart-treemap": "^0.17.15",
-    "@superset-ui/legacy-plugin-chart-world-map": "^0.17.15",
-    "@superset-ui/legacy-preset-chart-big-number": "^0.17.15",
+    "@superset-ui/legacy-plugin-chart-calendar": "^0.17.17",
+    "@superset-ui/legacy-plugin-chart-chord": "^0.17.17",
+    "@superset-ui/legacy-plugin-chart-country-map": "^0.17.17",
+    "@superset-ui/legacy-plugin-chart-event-flow": "^0.17.17",
+    "@superset-ui/legacy-plugin-chart-force-directed": "^0.17.17",
+    "@superset-ui/legacy-plugin-chart-heatmap": "^0.17.17",
+    "@superset-ui/legacy-plugin-chart-histogram": "^0.17.17",
+    "@superset-ui/legacy-plugin-chart-horizon": "^0.17.17",
+    "@superset-ui/legacy-plugin-chart-map-box": "^0.17.17",
+    "@superset-ui/legacy-plugin-chart-paired-t-test": "^0.17.17",
+    "@superset-ui/legacy-plugin-chart-parallel-coordinates": "^0.17.17",
+    "@superset-ui/legacy-plugin-chart-partition": "^0.17.17",
+    "@superset-ui/legacy-plugin-chart-pivot-table": "^0.17.17",
+    "@superset-ui/legacy-plugin-chart-rose": "^0.17.17",
+    "@superset-ui/legacy-plugin-chart-sankey": "^0.17.17",
+    "@superset-ui/legacy-plugin-chart-sankey-loop": "^0.17.17",
+    "@superset-ui/legacy-plugin-chart-sunburst": "^0.17.17",
+    "@superset-ui/legacy-plugin-chart-treemap": "^0.17.17",
+    "@superset-ui/legacy-plugin-chart-world-map": "^0.17.17",
+    "@superset-ui/legacy-preset-chart-big-number": "^0.17.17",
     "@superset-ui/legacy-preset-chart-deckgl": "^0.4.6",
-    "@superset-ui/legacy-preset-chart-nvd3": "^0.17.15",
-    "@superset-ui/plugin-chart-echarts": "^0.17.15",
-    "@superset-ui/plugin-chart-table": "^0.17.15",
-    "@superset-ui/plugin-chart-word-cloud": "^0.17.15",
-    "@superset-ui/preset-chart-xy": "^0.17.15",
+    "@superset-ui/legacy-preset-chart-nvd3": "^0.17.17",
+    "@superset-ui/plugin-chart-echarts": "^0.17.17",
+    "@superset-ui/plugin-chart-table": "^0.17.17",
+    "@superset-ui/plugin-chart-word-cloud": "^0.17.17",
+    "@superset-ui/preset-chart-xy": "^0.17.17",
     "@vx/responsive": "^0.0.195",
     "abortcontroller-polyfill": "^1.1.9",
     "antd": "^4.9.4",
